port: move the linux/osx console settings to src/port
