+++ include/port/console.h
@@ -1,0 +1,10 @@
+//
+// Created by hanyuan on 2024/4/23.
+//
+
+#ifndef TEMU_CONSOLE_H
+#define TEMU_CONSOLE_H
+
+void port_os_console_init();
+
+#endif //TEMU_CONSOLE_H
